Updated pom.xml to support Maven dependency management; moved source files to locations (probably) more suitable for development.
